// FE.Core.FW/src/apis/models/RegisteredCandidatesModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type RegisteredCandidatesModel = {
  id?: string;
  codeProfile?: string | null;
  userProfileId?: string;
  examPurpose?: string | null;
  scoreGoal?: number;
  isTested?: boolean;
  testDate?: string | null;
  placeOfRegistration?: string | null;
  examVersion?: string | null;
  submissionTime?: string;
  examId?: string;
  userId?: string | null;
  status?: number | null;
  statusPaid?: number | null;
  price?: number | null;
  areaId?: string | null;
};